destructure project page
