+++ components/results/ItemComparison.vue
@@ -1,0 +1,88 @@
+<template>
+  <ul class="item_comparison">
+    <li
+      v-for="(item, index) of items"
+      :key="index"
+      class="item_box"
+      :class="[
+        `item_${index + 1}`,
+        { active: activeId === item.id && activeSort },
+      ]"
+      @click="$emit('select', item.id)"
+    >
+      <font-awesome-icon
+        class="info"
+        icon="info-circle"
+        @click="$emit('showModal', item.id)"
+      />
+      {{ item.info.symbol }}
+
+      <font-awesome-icon
+        v-if="activeId === item.id && activeSort"
+        :icon="`sort-amount-${activeSort}`"
+      />
+    </li>
+  </ul>
+</template>
+
+<script>
+  export default {
+    props: {
+      items: {
+        type: Array,
+        default: () => [],
+      },
+      activeId: {
+        type: String,
+        default: '',
+      },
+      activeSort: {
+        type: String,
+        default: '',
+      },
+    },
+  };
+</script>
+
+<style lang="sass" scoped>
+  .item_comparison
+    padding: 0
+    margin: 0
+    display: flex
+    font-size: 16px
+    > .item_box
+      width: 100px
+      padding: 6px 0 0 0
+      font-weight: 500
+      margin-right: 10px
+      border-top: 8px solid
+      transition: width .3s
+      &.active
+        width: 200px
+        font-weight: bold
+      > svg
+        font-size: 14px
+        color: $MAIN_COLOR
+      &:hover
+        cursor: pointer
+      &.item_1
+        border-color: $COLOR_1
+      &.item_2
+        border-color: $COLOR_2
+      &.item_3
+        border-color: $COLOR_3
+      &.item_4
+        border-color: $COLOR_4
+      &.item_5
+        border-color: $COLOR_5
+      &.item_6
+        border-color: $COLOR_6
+      &.item_7
+        border-color: $COLOR_7
+      &.item_8
+        border-color: $COLOR_8
+      &.item_9
+        border-color: $COLOR_9
+      &.item_10
+        border-color: $COLOR_10
+</style>
